+++ core/math/camera_matrix.h
@@ -31,8 +31,14 @@
 #ifndef CAMERA_MATRIX_H
 #define CAMERA_MATRIX_H
 
-#include "core/math/rect2.h"
-#include "core/math/transform_3d.h"
+#include "core/math/math_defs.h"
+#include "core/math/vector3.h"
+
+class AABB;
+class Plane;
+class Transform3D;
+struct Rect2;
+struct Vector2;
 
 struct CameraMatrix {
 	enum Planes {
